fix dayccle events, add basic gamerules

+++ Content.Shared/_CP14/DayCycle/CP14DayCycleComponent.cs
@@ -4,8 +4,8 @@ namespace Content.Shared._CP14.DayCycle;
 [RegisterComponent]
 public sealed partial class CP14DayCycleComponent : Component
 {
-    public double LastLightLevel = 0f;
+    public float LastLightLevel = 0f;
 
     [DataField]
-    public double Threshold = 0.5f;
+    public float Threshold = 0.5f;
 }
